Update tests for session functions
