+++ backend/Makefile
@@ -3,7 +3,6 @@ build:
 
 test:
 	@go test github.com/djhranicky/ConcertTracker-SE-Project/routes
-	@go test github.com/djhranicky/ConcertTracker-SE-Project/service/auth
 
 run: build
 	@./bin/GO_BACKEND
